+++ CHANGES.rst
@@ -40,6 +40,8 @@ Development
 - Improve sandbox tooling incl. FTP source code upload
 - DS18B20: Extend time between starting the
   conversion and reading the sensor to one second.
+- Add MiniNet helper and corresponding ``Makefile`` rule ``wifi-connect``
+
 
 2019-06-22 0.5.1
 ================